requirements.txt
ffmpeg-python >= 0.2.0
nltk >= 3.9.1
openai-whisper >= 20240930
pydub >= 0.25.1
python-docx >= 1.1.2
streamlit >= 1.42.0
torch >= 2.6.0
datasets >= 3.3.2
transformers >= 4.44.2
